+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 - Release cray-nls 1.4.25 to update image used in iufBase.template.yaml (CASMINST-5621)
+- Built pre-install-toolkit, kubernetes, storage-ceph without changes
 - Release platform-utils v1.4.3 to edit output from ncnHealthChecks.sh (CASMTRIAGE-4539)
 - Released cray-nls v1.4.8 to add configmap for iuf-install-workflow-files (CASMINST-5530)
 - Released cray-keycloak-users-localize v1.11.3 to fix keycloak localize never ending (CASMTRIAGE-4286)